requirements.txt
alembic==1.13.3
annotated-types==0.7.0
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
Django==5.1.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.3
django-model-utils==5.0.0
django-notifications-hq==1.8.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
et-xmlfile==1.1.0
Faker==28.0.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonfield==3.1.0
linkify-it-py==2.0.3
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-py-plugins==0.4.1
mdurl==0.1.2
mysql-connector-python==9.0.0
mysqlclient==2.2.5
numpy==2.0.1
openpyxl==3.2.0b1
packaging==24.2
pandas==2.2.2
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
rich==13.7.1
six==1.16.0
SQLAlchemy==2.0.32
sqlparse==0.5.1
swapper==1.4.0
tabulate==0.9.0
textual==0.76.0
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
uritemplate==4.1.1
views==0.3
Werkzeug==3.0.4
